% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{data_deviation}
\alias{data_deviation}
\title{Deviation of Input Data from Euler Pole Solution}
\usage{
data_deviation(x, ep, sm = TRUE)
}
\arguments{
\item{x}{either a \code{matrix} or \code{data.frame} containing the geographical coordinates,
or a \code{sf} object.}

\item{ep}{three column vector. Euler pole solution}

\item{sm}{logical. Whether the solution is for a small circle or great circle geometries.}
}
\value{
numeric (angle in degrees). Angular difference of data points to the small or great circle.
}
\description{
Calculates angular distance of the input points to the small or great circle
of the Euler pole solution
}
\examples{
data(tintina)
res1 <- euler_solution(tintina)
data_deviation(tintina, res1)

data(south_atlantic)
res2 <- euler_solution(south_atlantic)
data_deviation(south_atlantic, res2)
}
