% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_solution.R
\name{data_misfit}
\alias{data_misfit}
\title{Misfit of input data from Euler pole solution}
\usage{
data_misfit(x, ep, sm = TRUE)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame} containing the geographical coordinates,
or a \code{sf} object}

\item{ep}{three column vector. Euler pole solution}

\item{sm}{logical.}
}
\value{
numeric (angle in degrees)
}
\description{
Calculates angular distance of the input points to the small or great circle
of the Euler pole solution
}
\examples{
data(tintina)
res <- euler_solution(tintina)
data_misfit(tintina, res)
}
