% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_solution.R
\name{deviation_stats}
\alias{deviation_stats}
\title{Distribution of deviation angles}
\usage{
deviation_stats(x)
}
\arguments{
\item{x}{Misfit angles in degrees}
}
\value{
\code{data.frame} giving the mean misfit angle,
its standard deviation, the dispersion from \code{0} degrees, the 95\% confidence angle,
the minimum and maximum absolute angles,
and the V2 test results (alternative hypothesis is an unimodal distribution around \code{0} degrees)
}
\description{
returns the mean deviation, its standard deviation, the variance,
the dispersion from zero deviation, the 95\% confidence angle, the minimum
and maximum absolute deviation angles, and the test results of the Rayleigh
test against a specified mean (i.e. zero deviation).
}
\examples{
data(tintina)
res <- euler_solution(tintina)
m <- data_deviation(tintina, res)
deviation_stats(m)
}
