% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress.R
\name{ep_from_stress}
\alias{ep_from_stress}
\title{Best Euler Pole Solution for Shmax Data}
\usage{
ep_from_stress(x, iter = 3, fact = 2, ...)
}
\arguments{
\item{x}{\code{data.frame} containing the coordinates of the point(s)
(\code{lat}, \code{lon}), the direction of
\eqn{\sigma_{Hmax}}{SHmax} \code{azi} and its standard deviation
\code{unc} (optional)}

\item{iter}{positive integer. Number of iterations to refine the Euler pole solution}

\item{fact}{positive integer. Aggregation factor to increase number of cells in each direction (horizontally and vertically, applied in each iteration.}

\item{...}{optional arguments passed to \code{\link[=ep_stress_dispersion]{ep_stress_dispersion()}}}
}
\value{
\code{sf} POINT object
}
\description{
Iteratively refines the Euler pole solution for a given set of
\eqn{\sigma_{Hmax}}{SHmax} data by
densifying the grid and selecting the Euler pole with the lowest dispersion.
}
\examples{
\dontrun{
data("san_andreas", package = "tectonicr")
ep_from_stress(san_andreas, prd = 135)
}
}
