% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_solution.R
\name{ep_help}
\alias{ep_help}
\alias{ep_from_sc}
\alias{ep_from_gc}
\title{Helper functions for euler pole solution}
\usage{
ep_from_sc(x)

ep_from_gc(x)
}
\arguments{
\item{x}{matrix}
}
\description{
Helper functions for euler pole solution
}
