% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ep_stereo_crs}
\alias{ep_stereo_crs}
\title{Stereographic projection centered in Euler pole}
\usage{
ep_stereo_crs(x)
}
\arguments{
\item{x}{two-element vector or data.frame specifying the latitude and longitude of the projections center.}
}
\value{
\code{sf::crs()} object
}
\description{
Stereographic projection centered in Euler pole
}
\examples{
ep_stereo_crs(c(45, 10))
}
