% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_reco.R
\docType{package}
\name{euler.reco}
\alias{euler.reco}
\alias{_PACKAGE}
\alias{euler.reco-package}
\title{library(euler.reco)}
\description{
Algorithm to find and evaluate the Euler pole solution describing the
orientation of geological structures
}
\details{
A list of documented functions may be viewed by typing
\code{help(package='euler.reco')}. A documentation and demonstration is
shown in the the vignette by typing
\code{vignette('intro', package = "euler.reco")}
}
