% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress.R
\name{euler_stress_dispersion}
\alias{euler_stress_dispersion}
\title{Stress Dispersion for Given Euler Pole Coordinates}
\usage{
euler_stress_dispersion(x, grid, prd, prob = 0.75)
}
\arguments{
\item{x}{\code{data.frame} containing the coordinates of the point(s)
(\code{lat}, \code{lon}), the direction of
\eqn{\sigma_{Hmax}}{SHmax} \code{azi} and its standard deviation
\code{unc} (optional)}

\item{grid}{grid given as \code{sfc} object}

\item{prd}{numeric. The theoretical orientation of \eqn{\sigma_{Hmax}}{SHmax} in PoR coordinates}

\item{prob}{Upper quantile threshold for circular dispersion to remove outliers. Default is 0.75}
}
\value{
SpatRaster
}
\description{
Stress Dispersion for Given Euler Pole Coordinates
}
\examples{
data("san_andreas", package = "tectonicr")
euler_stress_dispersion(san_andreas, grid = latlon_grid(gridsize = 10), prd = 135)
}
