% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{euler_summary}
\alias{euler_summary}
\title{Distribution of deviation angles}
\usage{
euler_summary(x)
}
\arguments{
\item{x}{numeric. Misfit angles in degrees}
}
\value{
\code{data.frame} giving the mean misfit angle,
its standard deviation, the dispersion from \code{0} degrees, the 95\% confidence angle,
the minimum and maximum absolute angles,
and the V2 test results (alternative hypothesis is an unimodal distribution around \code{0} degrees)
}
\description{
Mean deviation, standard deviation, variance, dispersion from zero deviation,
the 95\% confidence angle, the minimum
and maximum absolute deviation angles, and the test results of the Rayleigh
test against a specified mean (i.e. zero deviation).
}
\examples{
data(tintina)
res <- euler_reco(tintina)
m <- euler_deviation(tintina, res)
euler_summary(m)
}
