% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress.R
\name{extract_best_ep_from_grid}
\alias{extract_best_ep_from_grid}
\title{Extract Best Euler Pole from Dispersion Grid}
\usage{
extract_best_ep_from_grid(x, antipode = FALSE)
}
\arguments{
\item{x}{SpatRaster}

\item{antipode}{logical. If \code{TRUE}, the antipode of the Euler pole is returned}
}
\value{
\code{sf} POINT object
}
\description{
Extract Best Euler Pole from Dispersion Grid
}
\examples{
data("san_andreas", package = "tectonicr")
euler_stress_dispersion(san_andreas, grid = latlon_grid(gridsize = 10), prd = 135) |>
  extract_best_ep_from_grid()
}
