% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_solution.R
\name{geo_distribution}
\alias{geo_distribution}
\title{Position statistics}
\usage{
geo_distribution(x)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame} containing the geographical coordinates,
or a \code{sf} object}
}
\value{
three-element vector giving the mean geographic location, variance, and standard deviation
}
\description{
Statistics on the distribution of geographic locations
}
