% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{geo_stats}
\alias{geo_stats}
\title{Distribution of spherical coordinates}
\usage{
geo_stats(x)
}
\arguments{
\item{x}{either a \code{matrix} or \code{data.frame} containing the geographical coordinates,
or a \code{sf} object.}
}
\value{
three-element vector giving the mean geographic location, the standard deviation and the variance.
}
\description{
Statistics on the distribution of geographic locations
}
\examples{
xy <- cbind(lat=c(12,14,16), lon=c(-179,179, 177))
geo_stats(xy)
}
