% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{latlon_grid}
\alias{latlon_grid}
\title{Generates a grid from max and min coordinates}
\usage{
latlon_grid(
  x_range = c(-180, 180),
  y_range = c(0, 90),
  gridsize = 10,
  crs = 4326,
  ...
)
}
\arguments{
\item{x_range, y_range}{two-column vector giving the minimum and maximum coordinates. Unit as implied by \code{crs}}

\item{gridsize}{size of grid points. Unit as implied by \code{crs}}

\item{crs}{Coordinate reference system. Should be parsed by \code{\link[sf:st_crs]{sf::st_crs()}}. Default is \code{4326}}

\item{...}{arguments passed to \code{\link[sf:st_make_grid]{sf::st_make_grid()}}}
}
\value{
\code{sf} object
}
\description{
Generates a grid from max and min coordinates
}
\examples{
latlon_grid()
}
