% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_solution.R
\name{misfit_stats}
\alias{misfit_stats}
\title{Distribution of misfit angles}
\usage{
misfit_stats(x)
}
\arguments{
\item{x}{Misfit angles in degrees}
}
\value{
three column vector giving the mean misfit angle,
its standard deviation, the dispersion, the 95\% confidence angle, and the minimum and maximum absolute angles
}
\description{
Distribution of misfit angles
}
\examples{
data(tintina)
res <- euler_solution(tintina)
m <- data_misfit(tintina, res)
misfit_stats(m)
}
