% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_solution.R
\name{quick_plot}
\alias{quick_plot}
\title{Plot the Euler pole solution}
\usage{
quick_plot(x, sm = TRUE, sigdig = 1, omerc = FALSE)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame} containing the geographical coordinates,
or a \code{sf} object}

\item{sm}{logical. Whether the structure described by the points \code{x} is
expected to follow small (\code{TRUE}) or great circle (\code{FALSE}) arcs?}

\item{sigdig}{integer indicating the number of decimal places to be used.}

\item{omerc}{logical. Whether the plot should be shown in the oblique
Mercator projection with the Euler pole at North (\code{TRUE}) or not (\code{FALSE}, the default).}
}
\value{
ggplot
}
\description{
Plot the Euler pole solution
}
\examples{
data(tintina)
quick_plot(tintina)
quick_plot(tintina, omerc = TRUE)

data(south_atlantic)
quick_plot(south_atlantic)
quick_plot(south_atlantic, omerc = TRUE)
}
