% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_solution.R
\name{smallcircle}
\alias{smallcircle}
\title{Create a specified small circle around a given pole}
\usage{
smallcircle(lat, lon, angle = 90, n = 1000L)
}
\arguments{
\item{lat}{numeric. Latitude of the Euler pole (in degrees)}

\item{lon}{numeric. Longitude of the Euler pole  (in degrees)}

\item{angle}{numeric. Angle of small circle (if \code{angle = 90}, the great circle is returned)}

\item{n}{resolution of the return. Amount of points along small circle
(1000 by default).}
}
\value{
\code{sf} object
}
\description{
Create a specified small circle around a given pole
}
\examples{
smallcircle(45, 10, 12.3)
}
