% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{south_atlantic}
\alias{south_atlantic}
\title{South Atlantic}
\format{
An object of class \code{sf}
}
\usage{
data('south_atlantic')
}
\description{
example data set of a transform fault crossing the South Atlantic ridge
}
\examples{
data("south_atlantic")
head("south_atlantic")
}
\keyword{datasets}
