% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{tintina}
\alias{tintina}
\title{Tintina fault}
\format{
An object of class \code{sf}
}
\usage{
data('tintina')
}
\description{
example data set
}
\examples{
data("tintina")
head("tintina")
}
\keyword{datasets}
