% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{to_geomat}
\alias{to_geomat}
\title{Geographical coordinate matrix}
\usage{
to_geomat(x)
}
\arguments{
\item{x}{either a \code{matrix} or \code{data.frame} containing the geographical coordinates,
or a \code{sf} object.}
}
\value{
\code{matrix} of transformed coordinates
}
\description{
Creates the coordinates in matrix format
}
