% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{coordinates}
\alias{coordinates}
\alias{geographical_to_cartesian2}
\alias{cartesian_to_geographical2}
\title{Cartesian and geographic coordinates}
\usage{
geographical_to_cartesian2(x)

cartesian_to_geographical2(x)
}
\arguments{
\item{x}{\code{matrix} containing the coordinates}
}
\value{
\code{matrix} of the transformed coordinates
}
\description{
Convert between the two coordinate systems
}
