% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_solution.R
\name{coordinates}
\alias{coordinates}
\alias{geographical_to_cartesian2}
\alias{cartesian_to_geographical2}
\title{Coordinate conversion}
\usage{
geographical_to_cartesian2(x)

cartesian_to_geographical2(x)
}
\arguments{
\item{x}{\code{matrix}}
}
\value{
\code{matrix}
}
\description{
Coordinate conversion
}
