% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_solution.R
\name{quick_plot}
\alias{quick_plot}
\title{Plot the Euler pole solution}
\usage{
quick_plot(x, sm = TRUE, sigdig = 1, ...)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame} containing the geographical coordinates,
or a \code{sf} object}

\item{sm}{logical. Whether the structure described by the points \code{x} is
expected to follow small (\code{TRUE}) or great circle (\code{FALSE}) arcs?}

\item{sigdig}{integer indicating the number of decimal places (round) to be used.}

\item{...}{optional arguments passed to \code{\link[=euler_solution]{euler_solution()}}}
}
\value{
ggplot
}
\description{
Plot the Euler pole solution
}
\examples{
data(tintina)
quick_plot(tintina)
}
