% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{quick_plot}
\alias{quick_plot}
\title{Plot the Euler pole solution}
\usage{
quick_plot(
  x,
  sc = TRUE,
  densify.x = FALSE,
  ...,
  proj = c("geo", "omerc", "stereo"),
  expand = c(0, 0)
)
}
\arguments{
\item{x}{either a \code{matrix} or \code{data.frame} containing the geographical coordinates,
or a \code{sf} object.}

\item{sc}{logical. Whether the structure described by the points \code{x} is
expected to follow small (\code{TRUE}) or great circle (\code{FALSE}) arcs?}

\item{densify.x}{logical. Whether the points \code{x} along the lines structure
should be densified before analysis (\code{TRUE}) or not (\code{FALSE}, the default).
Densification would yield equally spaced points along the line.}

\item{...}{optional arguments passed to \code{\link[smoothr:densify]{smoothr::densify()}} (only if \code{densify.x = TRUE}).}

\item{proj}{character. Whether the plot should be shown in the geographic Mercator projection (\code{"geo"}, the default), the oblique
Mercator projection with the Euler pole at North (\code{"omerc"}), or the stereographic projection centered in the Euler pole (\code{"stereo"}).}

\item{expand}{numeric two element vector.  expand the map limits in latitude and longitude (\code{c(0, 0)} (degrees) by default)}
}
\description{
Creates a simple ggplot to show the best fit small/great circle,
the data's misfit to the solution and some statistics of the solution.
}
\examples{
data(tintina)
quick_plot(tintina)
quick_plot(tintina, proj = "omerc")

data(south_atlantic)
quick_plot(south_atlantic)
quick_plot(south_atlantic, proj = "omerc")
quick_plot(south_atlantic, proj = "stereo")
}
