% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{smallcircle}
\alias{smallcircle}
\title{#' @rdname coordinates
geographical_to_acoscartesian <- function(x) {
structr:::cartesian_to_acoscartesian(
geographical_to_cartesian2(x)
)
}}
\usage{
smallcircle(lat, lon, angle = 90, n = 1000L)
}
\arguments{
\item{lat}{numeric. Latitude of the Euler pole (in degrees)}

\item{lon}{numeric. Longitude of the Euler pole  (in degrees)}

\item{angle}{numeric. Angle of small circle (half apical angle of the spherical cone).
If \code{angle = 90}, the great circle is returned.}

\item{n}{integer. resolution of the return. Amount of points along small circle
(1000 by default).}
}
\value{
\code{sf} object
}
\description{
Creates a small circle (or spherical cone) by specifying the axis and the half apical angle
}
\details{
#' @rdname coordinates
acoscartesian_to_geographical <- function(x) {
cartesian_to_geographical2(
structr:::acoscartesian_to_cartesian(x)
)
}

Small circle around a given pole
}
\examples{
smallcircle(45, 10, 12.3)
}
