% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_solution.R
\name{to_geomat}
\alias{to_geomat}
\title{Convert object to geographic coordinate matrix}
\usage{
to_geomat(x)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame} containing the geographical coordinates,
or a \code{sf} object}
}
\value{
\code{matrix}
}
\description{
Convert object to geographic coordinate matrix
}
